удаляет favorities card
