+++ src/components/card/card.tsx
@@ -25,7 +25,7 @@ const Card = ({offer, block}: CardProps): JSX.Element => {
         <div className="place-card__mark">
           <span>Premium</span>
         </div>)}
-      <div className="cities__image-wrapper place-card__image-wrapper">
+      <div className={`${block}__image-wrapper place-card__image-wrapper`}>
         <Link to={`${AppRoute.Offer}/${id}`}>
           <img className="place-card__image" src={previewImage} width="260" height="200" alt={title}></img>
         </Link>
