--- src/components/favorites-card/favorities-card.tsx
@@ -1,57 +1,0 @@
-import { OfferPreview } from '../../types/offers-preview';
-import { Link } from 'react-router-dom';
-import { getRatingWidth } from '../../utils/cards';
-import { AppRoute } from '../../const/const';
-
-type FavoritiesCardProps = {
-  offer: OfferPreview;
-}
-
-const FavoritiesCard = ({offer}: FavoritiesCardProps): JSX.Element | null => {
-
-  const {isPremium, previewImage, price, rating, title, type, id, isFavorite } = offer;
-
-  if(!isFavorite){
-    return null;
-  }
-  return(
-    <article className="favorites__card place-card">
-      {isPremium && (
-        <div className="place-card__mark">
-          <span>Premium</span>
-        </div>)}
-      <div className="favorites__image-wrapper place-card__image-wrapper">
-        <Link to={`${AppRoute.Offer}/${id}`}>
-          <img className="place-card__image" src={previewImage} width="150" height="110" alt="Place image"></img>
-        </Link>
-      </div>
-      <div className="favorites__card-info place-card__info">
-        <div className="place-card__price-wrapper">
-          <div className="place-card__price">
-            <b className="place-card__price-value">&euro;{price}</b>
-            <span className="place-card__price-text">&#47;&nbsp;night</span>
-          </div>
-          <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
-            <svg className="place-card__bookmark-icon" width="18" height="19">
-              <use xlinkHref="#icon-bookmark"></use>
-            </svg>
-            <span className="visually-hidden">In bookmarks</span>
-          </button>
-        </div>
-        <div className="place-card__rating rating">
-          <div className="place-card__stars rating__stars">
-            <span style={{width: getRatingWidth(rating)}}></span>
-            <span className="visually-hidden">Rating</span>
-          </div>
-        </div>
-        <h2 className="place-card__name">
-          <Link to={`${AppRoute.Offer}/${id}`}>{title}</Link>
-        </h2>
-        <p className="place-card__type">{type}</p>
-      </div>
-    </article>
-  );
-};
-
-
-export default FavoritiesCard;
